services/treasury_strats/requirements.txt
fastapi==0.115.14 
uvicorn[standard]==0.35.0 
pydantic==2.11.7 
pydantic-settings==2.10.1 
web3==7.5.0 
pyyaml==6.0.2 
pandas==2.2.3 
apscheduler==3.10.4 
prometheus-client==0.22.1 
python-dateutil==2.9.0.post0 
 
